Revert "Split files by method and execution to avoid duplication of test and program"
This reverts commit 6746ba5780a14f52b76c4b39cb9d81397335e280.
