+++ 15.ls_again/test/ls_test.rb
@@ -1,10 +1,9 @@
 # frozen_string_literal: true
 
 require 'minitest/autorun'
-require_relative '../lib/ls_def'
+require_relative '../lib/ls'
 
 TEST_DIR = File.join(__dir__, 'ls-sample')
-COLUMNS = 3
 
 class LsTest < Minitest::Test
   def test_1_files
